Add github deploy action
