+++ .github/workflows/main.yml
@@ -1,0 +1,46 @@
+name: Teamo Master
+
+on:
+    push:
+        branches: [ feature/github-actions ]
+
+jobs:
+  deploy:
+    name: Deploy
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v2
+    - name: Clean up Teamo previous deploy
+      uses: appleboy/ssh-action@master
+      with:
+        host: ${{ secrets.DEPLOY_HOST }}
+        username: ${{ secrets.DEPLOY_USERNAME }}
+        passphrase: ${{ secrets.DEPLOY_PASSPHRASE }}
+        key: ${{ secrets.DEPLOY_KEY }}
+        port: ${{ secrets.DEPLOY_PORT }}
+        script: |
+          rm -r teamo/master/src
+          docker stop teamo-master || true
+          docker rm teamo-master || true
+    - name: Copy Teamo to server
+      uses: appleboy/scp-action@master
+      with:
+        host: ${{ secrets.DEPLOY_HOST }}
+        username: ${{ secrets.DEPLOY_USERNAME }}
+        passphrase: ${{ secrets.DEPLOY_PASSPHRASE }}
+        key: ${{ secrets.DEPLOY_KEY }}
+        port: ${{ secrets.DEPLOY_PORT }}
+        source: '*/*.*, *.*'
+        target: "teamo/master/src"
+    - name: Start Docker container
+      uses: appleboy/ssh-action@master
+      with:
+        host: ${{ secrets.DEPLOY_HOST }}
+        username: ${{ secrets.DEPLOY_USERNAME }}
+        passphrase: ${{ secrets.DEPLOY_PASSPHRASE }}
+        key: ${{ secrets.DEPLOY_KEY }}
+        port: ${{ secrets.DEPLOY_PORT }}
+        script: |
+          docker build -t teamo-master teamo/master/src/
+          python teamo/master/src/create_database.py teamo/master/db/teamo.db
+          docker run -e TEAMO_BOT_TOKEN=${{ secrets.TEAMO_BOT_TOKEN_MASTER }} -d -v $PWD/teamo/master/db/teamo.db:/teamo/db/teamo.db
